correct entry point initialization
